+++ game.lua
@@ -1,26 +1,43 @@
-local angle = 360
-local scale = player.scale
-local params = vector:new(player.w, player.h)
-local pos = player.pos
+local scale = vector:new(10, 10)
+local pos = vector:new(100, 160)
+local params
+local p1
 
-do_imgui = function()
-    if imgui:Begin(window, "Properties", 300, 300) then
-        if imgui:begin_section(window, "Transform") then
+function init(win)
+    p1 = player:new(win)
+    p1.angle = 1
+    p1.scale = scale
+    p1.pos = pos
 
-            if imgui:hslider_float(window, "Rotation", 0, 720, angle) then
-                player.angle = angle
+    params = vector:new(p1.w, p1.h)
+
+    imgui.init(win)
+end
+
+function update(dt)
+    p1:update(dt)
+    p1:rotate(dt/100)
+end
+
+function draw(win)
+    p1:draw(win)
+end
+
+
+function do_imgui(win)
+    if imgui.Begin(win, "Properties", 300, 300) then
+        if imgui.begin_section(win, "Transform") then
+            if imgui.hslider_vec(win, "Scale", 0.5, 50, 0.5, 50, scale) then
+                p1.scale = scale
             end
-            if imgui:hslider_vec(window, "Scale", 0.5, 50, 0.5, 50, scale) then
-                player.scale = scale
-            end
-            if imgui:hslider_vec(window, "Position",
+            if imgui.hslider_vec(win, "Position",
                     params.x/2, screen_w - (params.x/2),
                     params.y/2, screen_h - (params.y/2), pos) then
-                player.pos = pos
+                p1.pos = pos
             end
 
-            imgui:end_section()
+            imgui.end_section()
         end
-        imgui:End()
+        imgui.End()
     end
 end