+++ src/fileIO/Snipper.cpp
@@ -3,44 +3,35 @@
 
 #include "fileIO/Snipper.hpp"
 
-std::recursive_mutex tnt::file::Snipper::mtx;
+namespace fs = std::filesystem;
 
-tnt::file::Snipper::~Snipper() noexcept
+void tnt::Snipper::watchFile(std::string_view name)
 {
     std::lock_guard lock{mtx};
-    for (auto f : files)
-        if (f.second != nullptr)
-        {
-            delete files[f.first];
-            files[f.first] = nullptr;
-        }
-    files.clear();
-    std::map<std::string, File *>{}.swap(files);
-}
-
-void tnt::file::Snipper::watchFile(std::string_view name)
-{
-    std::lock_guard lock{mtx};
-    if (files[name.data()])
+    std::string key{name.data()};
+    if (fs::exists(key))
         return;
-    files[name.data()] = new File{name};
+    files.try_emplace(key, key, fs::last_write_time(key));
 }
 
-bool tnt::file::Snipper::isModified(std::string_view file) noexcept
+bool tnt::Snipper::isModified(std::string_view file) noexcept
 {
     std::lock_guard lock{mtx};
-    if (auto pos{files.find(file.data())}; pos == files.end())
-        files[file.data()] = new File{file};
-    if (auto time{std::filesystem::last_write_time(std::filesystem::path{file})}; time != files[file.data()]->lastTime)
+    std::string key{file.data()};
+    if (auto pos{files.find(key)}; pos == files.cend())
+        files.try_emplace(key, key, fs::last_write_time(key));
+
+    if (auto time{fs::last_write_time(key)};
+        time != files[key].second)
     {
-        files[file.data()]->lastTime = time;
+        files[key].second = time;
         return true;
     }
     return false;
 }
 
-void tnt::file::Snipper::unwatchFile(std::string_view filename) noexcept
+void tnt::Snipper::unwatchFile(std::string_view filename) noexcept
 {
-    if (auto it{files.find(filename.data())}; it != files.end())
+    if (auto it{files.find(filename.data())}; it != files.cend())
         it = files.erase(it);
 }